add loadout slots validation test and fix dclass gloves loadout
