+++ Content.Common/Inventory/SlotFlags.cs
@@ -1,5 +1,3 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
 using Robust.Shared.Serialization;
 
 namespace Content.Shared.Inventory;
