+++ Content.IntegrationTests/Content.IntegrationTests.csproj
@@ -11,17 +11,19 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Content.Client\Content.Client.csproj" />
-    <ProjectReference Include="..\Content.Goobstation.Client\Content.Goobstation.Client.csproj" />
-    <ProjectReference Include="..\Content.Goobstation.Server\Content.Goobstation.Server.csproj" />
-    <ProjectReference Include="..\Content.Goobstation.Shared\Content.Goobstation.Shared.csproj" />
     <!-- Trauma -->
+    <ProjectReference Include="../Content.Common/Content.Common.csproj" />
+    <ProjectReference Include="../Content.Goobstation.Client/Content.Goobstation.Client.csproj" />
+    <ProjectReference Include="../Content.Goobstation.Server/Content.Goobstation.Server.csproj" />
+    <ProjectReference Include="../Content.Goobstation.Shared/Content.Goobstation.Shared.csproj" />
+    <ProjectReference Include="../Content.Trauma.Client/Content.Trauma.Client.csproj" />
     <ProjectReference Include="../Content.Trauma.Client/Content.Trauma.Client.csproj" />
     <ProjectReference Include="../Content.Trauma.Server/Content.Trauma.Server.csproj" />
     <ProjectReference Include="../Content.Trauma.Shared/Content.Trauma.Shared.csproj" />
     <ProjectReference Include="../Content.Lavaland.Shared/Content.Lavaland.Shared.csproj" />
     <ProjectReference Include="../Content.Medical.Shared/Content.Medical.Shared.csproj" />
     <!-- /Trauma -->
+    <ProjectReference Include="..\Content.Client\Content.Client.csproj" />
     <ProjectReference Include="..\Content.Server\Content.Server.csproj" />
     <ProjectReference Include="..\Content.Shared\Content.Shared.csproj" />
     <ProjectReference Include="..\Content.Tests\Content.Tests.csproj" />
