+++ Content.Trauma.Common/Inventory/InventorySlotPrototype.cs
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+using Content.Shared.Inventory;
+
 namespace Content.Trauma.Common.Inventory;
 
 /// <summary>
@@ -11,4 +13,10 @@ public sealed partial class InventorySlotPrototype : IPrototype
 {
     [IdDataField]
     public string ID { get; private set; } = default!;
+
+    /// <summary>
+    /// Slot flags that correspond to this slot.
+    /// </summary>
+    [DataField]
+    public SlotFlags Flags = SlotFlags.NONE;
 }
